+++ src/main/java/myapp/SzakdolgozatBE/genre/GenreService.java
@@ -11,7 +11,12 @@ public class GenreService {
     GenreDAO dao;
 
     public Genre getGenre(long id) throws NullPointerException {
-        return dao.getGenre(id);
+        Genre tmp = dao.getGenre(id);
+        if (tmp == null) {
+            throw new NullPointerException();
+        } else {
+            return tmp;
+        }
     }
 
     public List<Genre> getAllGenres() {
@@ -25,12 +30,21 @@ public class GenreService {
     }
 
     public void deleteGenre(long id) {
-        dao.deleteGenre(id);
+        Genre tmp = dao.getGenre(id);
+        if (tmp == null) {
+            throw new NullPointerException();
+        } else {
+            dao.deleteGenre(id);
+        }
     }
 
     public Genre modifyGenre(long id, String name) throws NullPointerException {
-        Genre tmp = new Genre();
-        tmp.setName(name);
-        return dao.modifyGenre(id, tmp);
+        Genre tmp = dao.getGenre(id);
+        if (tmp != null) {
+            tmp.setName(name);            
+            return dao.modifyGenre(tmp);
+        } else {
+            throw new NullPointerException();
+        }
     }
 }
