+++ src/main/java/myapp/SzakdolgozatBE/movie/MovieService.java
@@ -17,25 +17,18 @@ public class MovieService {
     
     @Inject
     GenreService genreService;
-
-    /*public Movie addMovie(String title, int budget, int length, int releaseYear, byte[] coverPicture, long ageClassificationId, long genreId) {
-        Movie tmp = new Movie();
-        tmp.setTitle(title);
-        tmp.setBudget(budget);
-        tmp.setLength(length);
-        tmp.setReleaseYear(releaseYear);
-        tmp.setCoverPicture(coverPicture);
-        tmp.setAgeClassification(ageClassificationService.getAgeClassification(ageClassificationId));
-        tmp.setGenre(genreService.getGenre(genreId));
-        return dao.addMovie(tmp);
-    }*/
     
     public Movie addMovie(Movie movie) {
         return dao.addMovie(movie);
     }
 
     public Movie getMovie(long id) throws NullPointerException {
-        return dao.getMovie(id);
+        Movie tmp = dao.getMovie(id);
+        if (tmp == null) {
+            throw new NullPointerException();
+        } else {
+            return tmp;
+        }
     }
 
     public List<Movie> getAllMovies() {
@@ -43,19 +36,29 @@ public class MovieService {
     }
 
     public void deleteMovie(long id) throws NullPointerException {
-        dao.deleteMovie(id);
+        Movie tmp = dao.getMovie(id);
+        if (tmp == null) {
+            throw new NullPointerException();
+        } else {
+            dao.deleteMovie(id);
+        }
     }
 
-    public Movie modifyMovie(long id, String title, int budget, int length, int releaseYear, byte[] coverPicture, long ageClassificationId, long genreId) throws NullPointerException {
-        Movie tmp = new Movie();
-        tmp.setTitle(title);
-        tmp.setBudget(budget);
-        tmp.setLength(length);
-        tmp.setReleaseYear(releaseYear);
-        tmp.setCoverPicture(coverPicture);
-        tmp.setAgeClassification(ageClassificationService.getAgeClassification(ageClassificationId));
-        tmp.setGenre(genreService.getGenre(genreId));
-        return dao.modifyMovie(id, tmp);
+    public Movie modifyMovie(long id, Movie movie) throws NullPointerException {
+        Movie tmp = dao.getMovie(id);
+        if (tmp != null) {
+            tmp.setTitle(movie.getTitle());
+            tmp.setBudget(movie.getBudget());
+            tmp.setLength(movie.getLength());
+            tmp.setReleaseYear(movie.getReleaseYear());
+            tmp.setCoverPicture(movie.getCoverPicture());
+            tmp.setAgeClassification(movie.getAgeClassification());
+            tmp.setGenre(movie.getGenre());
+            return dao.modifyMovie(tmp);
+        } else {
+            throw new NullPointerException();
+        }
+        
     }
 
     public void changeRating(long id, int rating) {
