+++ src/main/java/myapp/SzakdolgozatBE/season/SeasonSerivce.java
@@ -3,6 +3,7 @@ package myapp.SzakdolgozatBE.season;
 import java.util.List;
 import javax.ejb.Stateless;
 import javax.inject.Inject;
+import myapp.SzakdolgozatBE.series.Series;
 import myapp.SzakdolgozatBE.series.SeriesDAO;
 
 @Stateless
@@ -14,29 +15,50 @@ public class SeasonSerivce {
     @Inject
     SeriesDAO seriesDAO;
 
-    public Season addSeason(int number, long seriesId) {
-        Season tmp = new Season();
-        tmp.setNumber(number);
-        tmp.setSeries(seriesDAO.getSeries(seriesId));
-        return dao.addSeason(tmp);
+    public Season addSeason(Season season) throws NullPointerException {
+        Series tmp = seriesDAO.getSeries(season.getSeries().getId());
+        if (tmp != null) {            
+            return dao.addSeason(season);
+        } else {
+            throw new NullPointerException();
+        }
     }
 
     public Season getSeason(long id) throws NullPointerException {
-        return dao.getSeason(id);
+        Season tmp = dao.getSeason(id);
+        if (tmp == null) {
+            throw new NullPointerException();
+        } else {
+            return tmp;
+        }
     }
 
-    public List<Season> getSeriesSeasons(long seriesId) {
-        return dao.getSeriesSeasons(seriesDAO.getSeries(seriesId));
+    public List<Season> getSeriesSeasons(long seriesId) throws NullPointerException {
+        Series tmp = seriesDAO.getSeries(seriesId);
+        if (tmp != null) {
+            return dao.getSeriesSeasons(tmp);
+        } else {
+            throw new NullPointerException();
+        }
     }
 
     public void deleteSeason(long id) throws NullPointerException {
-        dao.deleteSeason(id);
+        Season tmp = dao.getSeason(id);
+        if (tmp != null) {
+            dao.deleteSeason(id);
+        } else {
+            throw new NullPointerException();
+        }
     }
 
-    public Season modifySeason(long id, int number, long seriesId) throws NullPointerException {
-        Season tmp = new Season();
-        tmp.setNumber(number);
-        tmp.setSeries(seriesDAO.getSeries(seriesId));
-        return dao.modifySeason(id, tmp);
+    public Season modifySeason(long id, Season season) throws NullPointerException {
+        Season tmp = dao.getSeason(id);
+        if (tmp != null) {
+            tmp.setNumber(season.getNumber());
+            tmp.setSeries(season.getSeries());
+            return dao.modifySeason(tmp);
+        } else {
+            throw new NullPointerException();
+        }
     }
 }
