+++ src/main/java/myapp/SzakdolgozatBE/series/SeriesService.java
@@ -18,18 +18,17 @@ public class SeriesService {
     @Inject
     GenreService genreService;
 
-    public Series addSeries(String title, int releaseYear, byte[] coverPicture, long ageClassificationId, long genreId) {
-        Series tmp = new Series();
-        tmp.setTitle(title);
-        tmp.setReleaseYear(releaseYear);
-        tmp.setCoverPicture(coverPicture);
-        tmp.setAgeClassification(ageClassificationService.getAgeClassification(ageClassificationId));
-        tmp.setGenre(genreService.getGenre(genreId));
-        return dao.addSeries(tmp);
+    public Series addSeries(Series series) {
+        return dao.addSeries(series);
     }
 
     public Series getSeries(long id) throws NullPointerException {
-        return dao.getSeries(id);
+        Series tmp = dao.getSeries(id);
+        if (tmp == null) {
+            throw new NullPointerException();
+        } else {
+            return tmp;
+        }
     }
 
     public List<Series> getAllSeries() {
@@ -37,17 +36,26 @@ public class SeriesService {
     }
 
     public void deleteSeries(long id) throws NullPointerException {
-        dao.deleteSeries(id);
+        Series tmp = dao.getSeries(id);
+        if (tmp != null) {
+            dao.deleteSeries(id);
+        } else {
+            throw new NullPointerException();
+        }
     }
 
-    public Series modifySeries(long id, String title, int releaseYear, byte[] coverPicture, long ageClassificationId, long genreId) throws NullPointerException {
-        Series tmp = new Series();
-        tmp.setTitle(title);
-        tmp.setReleaseYear(releaseYear);
-        tmp.setCoverPicture(coverPicture);
-        tmp.setAgeClassification(ageClassificationService.getAgeClassification(ageClassificationId));
-        tmp.setGenre(genreService.getGenre(genreId));
-        return dao.modifySeries(id, tmp);
+    public Series modifySeries(long id, Series series) throws NullPointerException {
+        Series tmp = dao.getSeries(id);
+        if (tmp != null) {
+            tmp.setTitle(series.getTitle());
+            tmp.setReleaseYear(series.getReleaseYear());
+            tmp.setCoverPicture(series.getCoverPicture());
+            tmp.setAgeClassification(series.getAgeClassification());
+            tmp.setGenre(series.getGenre());
+            return dao.modifySeries(tmp);
+        } else {
+            throw new NullPointerException();
+        }
     }
 
     public void changeRating(long id) {
